factor courseDetails into SingleCourse
