+++ backend/src/main/java/pl/edu/pw/pap/PapApplication.java
@@ -42,7 +42,7 @@ public class PapApplication {
         };
     }
 
-//        @Bean
+        @Bean
     @Profile({"dev", "dev-postgres"})
     public CommandLineRunner addDummyData(
             CourseRepository courseRepository,
